LayoutTests/platform/gtk/accessibility/w3c-svg-name-calculation-expected.txt

This tests the accessible text alternatives results for SVG.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


test1:
	Expected name: end
	AXTitle: end
	AXDescription: abc
test2:
	Expected name: end
	AXTitle: end
	AXDescription: 
test3:
	Expected name: end
	AXTitle: end
	AXDescription: abc
test4:
	Expected name: end
	AXTitle: end
	AXDescription: 
test5:
	Expected name: hello
	AXTitle: hello
	AXDescription: abc
test6:
	Expected name: hello
	AXTitle: hello
	AXDescription: 
test7:
	Expected name: abc
	AXTitle: abc
	AXDescription: 
test8:
	Expected name: hi
	AXTitle: hi
	AXDescription: 
test9:
	Expected name: hi
	AXTitle: hi
	AXDescription: 
test10:
	Expected name: hi
	AXTitle: hi
	AXDescription: 
test11:
	Expected name: hola
	AXTitle: hola
	AXDescription: 
test12:
	Expected name: hola
	AXTitle: hola
	AXDescription: 
test13:
	Expected name: hola
	AXTitle: hola
	AXDescription: 
test14:
	Expected name: bob
	AXTitle: bob
	AXDescription: abc
test15:
	Expected name: bob
	AXTitle: bob
	AXDescription: 
test16:
	Expected name: abc
	AXTitle: abc
	AXDescription: 
test17:
	Expected name: (empty)
	AXTitle: 
	AXDescription: 
test18:
	Expected name: waz up
	AXTitle: waz up
	AXDescription: 
test19:
	Expected name: Hallo
	AXTitle: Hallo
	AXDescription: 
test20:
	Expected name: the end
	AXTitle: the end
	AXDescription: 
test21:
	Expected name: big end
	AXTitle: big end
	AXDescription: 
test22:
	Expected name: big end
	AXTitle: big end
	AXDescription: 
test23:
	Expected name: end
	AXTitle: end
	AXDescription: 
test24:
	Expected name: end
	AXTitle: end
	AXDescription: 
test25:
	Expected name: booth
	AXTitle: booth
	AXDescription: 
test26:
	Expected name: table
	AXTitle: table
	AXDescription: 
test27:
	Expected name: counter
	AXTitle: counter
	AXDescription: 
test28:
	Expected name: counter
	AXTitle: counter
	AXDescription: 
test29:
	Expected name: counter
	AXTitle: counter
	AXDescription: 
test30:
	Expected name: boulder
	AXTitle: boulder
	AXDescription: 
test31:
	Expected name: stone
	AXTitle: stone
	AXDescription: 
test32:
	Expected name: pebble
	AXTitle: pebble
	AXDescription: 
test33:
	Expected name: rock
	AXTitle: rock
	AXDescription: 
test34:
	Expected name: rock
	AXTitle: rock
	AXDescription: 
test35:
	Expected name: stone
	AXTitle: stone
	AXDescription: 
test36:
	Expected name: stone
	AXTitle: stone
	AXDescription: 
test37:
	Expected name: piedra
	AXTitle: piedra
	AXDescription: 
test38:
	Expected name: piedra
	AXTitle: piedra
	AXDescription: 
test39:
	Expected name: stone
	AXTitle: stone
	AXDescription: 
test40:
	Expected name: stone
	AXTitle: stone
	AXDescription: 
test41:
	Expected name: (empty)
	AXTitle: 
	AXDescription: 
test42:
	Expected name: (empty)
	AXTitle:  
	AXDescription: 
test43:
	Expected name: (empty)
	AXTitle: 
	AXDescription: 
test44:
	Expected name: (empty)
	AXTitle:  
	AXDescription: 
test45:
	Expected name: (empty)
	AXTitle: 
	AXDescription:  
test46:
	Expected name: (empty)
	AXTitle:  
	AXDescription: 	
PASS successfullyParsed is true

TEST COMPLETE
